Convert medea to lib for docs.rs
